Create quantrautil.py
This file is used to fetch the data for different web sources. This code needs to be stored in the quantrautil.py. And save the file in .....\Anaconda2\Lib\site-packages folder or in your current working folder. You have to install the necessary packages and the line no 13 needs to be replaced with your Quandl key for the code to work correctly